updated console.log and format in foundry

+++ src/Counter.sol
@@ -1,19 +1,21 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity ^0.8.13;
 
+// import "forge-std/console.sol";
+
 contract Counter {
-    uint public count; 
-    
-    function get() public view returns (uint) {
+    uint256 public count;
+
+    function get() public view returns (uint256) {
         return count;
     }
 
-    function inc() public{
-        count++; 
+    function inc() public {
+        // console.log("HERE", count);
+        count++;
     }
-
-    function dec() public{ 
+    
+    function dec() public {
         count--;
     }
-
 }
